Move SignalMixin and TimeMixin back to common_util
This concludes our service.
